% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{apply_coin_op}
\alias{apply_coin_op}
\title{Apply a coin operator to a walk state.}
\usage{
apply_coin_op(state, coin_matrix)
}
\arguments{
\item{state}{represents the state matrix of the walk.}

\item{coin_matrix}{a \eqn{2 \times 2} matrix acting on the coin space.}
}
\value{
the updated state matrix after the coin operator is applied.
}
\description{
This applies a \eqn{2 \times 2} coin operator independently
at each position of the walk.
}
\examples{
state <- initialize_state(1, "right")
apply_coin_op(state, hadamard_coin())

}
