% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{apply_shift_op}
\alias{apply_shift_op}
\title{Apply the conditional shift operator.}
\usage{
apply_shift_op(state)
}
\arguments{
\item{state}{represents the state matrix of the walk.}
}
\value{
the updated state matrix after the conditional shift.
}
\description{
This applies the conditional shift operator for the walk.
The shift operator sends the right coin state one step to the
right and the left coin state one step to the left:
\deqn{
S|x,\mathrm{right}\rangle = |x+1,\mathrm{right}\rangle,
\qquad
S|x,\mathrm{left}\rangle = |x-1,\mathrm{left}\rangle.
}
}
\examples{
state <- initialize_state(1, "right")
apply_shift_op(state)

}
