% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{apply_unitary_step}
\alias{apply_unitary_step}
\title{Apply one noiseless Hadamard quantum walk step.}
\usage{
apply_unitary_step(state)
}
\arguments{
\item{state}{represents the state matrix of the walk.}
}
\value{
the updated state matrix after one noiseless walk step.
}
\description{
This applies one noiseless step of the Hadamard quantum random
walk. The unitary step is
\deqn{
U = S(H \otimes I_p),
}
where \eqn{H} is the Hadamard coin, \eqn{S} is the conditional shift, and
\eqn{I_p} is the identity on the position space.
}
\examples{
state <- initialize_state(2, "symmetric")
apply_unitary_step(state)

}
