% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{basis_left}
\alias{basis_left}
\title{Construct the left coin basis vector.}
\usage{
basis_left()
}
\value{
a \eqn{2 \times 1} column vector representing the left coin state.
}
\description{
This returns the left coin basis state as a column vector.
}
\examples{
basis_left()

}
