% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{basis_right}
\alias{basis_right}
\title{Construct the right coin basis vector.}
\usage{
basis_right()
}
\value{
a \eqn{2 \times 1} column vector representing the right coin state.
}
\description{
This returns the right coin basis state as a column vector.
}
\examples{
basis_right()

}
