% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{get_clean_numeric}
\alias{get_clean_numeric}
\title{Clean a numeric vector.}
\usage{
get_clean_numeric(data)
}
\arguments{
\item{data}{the vector to clean.}
}
\value{
A vector containing the cleaned data.
}
\description{
Cleans an inputted vector of decimal points, characters, and missing values.
}
\examples{
x <- c("1,000", "2,500", "", NA, 1)
get_clean_numeric(x)
}
