% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_kraus_ops}
\alias{get_kraus_ops}
\title{Select the Kraus operator family for a channel.}
\usage{
get_kraus_ops(channel, p)
}
\arguments{
\item{channel}{represents the selected decoherence channel.}

\item{p}{represents the decoherence strength.}
}
\value{
a list of Kraus operators for the selected channel.
}
\description{
This returns the appropriate Kraus operator family for either
the dephasing or depolarizing channel.
}
\examples{
get_kraus_ops("dephasing", 0.1)

}
