% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hadamard_coin}
\alias{hadamard_coin}
\title{Construct the Hadamard coin operator.}
\usage{
hadamard_coin()
}
\value{
a \eqn{2 \times 2} matrix representing the Hadamard coin.
}
\description{
This returns the \eqn{2 \times 2} Hadamard matrix acting on
the coin space. The Hadamard coin is
\deqn{
H = \frac{1}{\sqrt{2}}
\begin{pmatrix}
1 & 1 \\
1 & -1
\end{pmatrix}.
}
}
\examples{
hadamard_coin()

}
