% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{identity_coin}
\alias{identity_coin}
\title{Construct the identity operator on the coin space.}
\usage{
identity_coin()
}
\value{
a \eqn{2 \times 2} identity matrix.
}
\description{
This returns the identity matrix acting on the
two-dimensional coin space.
}
\examples{
identity_coin()

}
