% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{init_coin_state}
\alias{init_coin_state}
\title{Construct the initial coin state.}
\usage{
init_coin_state(init_coin)
}
\arguments{
\item{init_coin}{represents the initial coin state used to start the walk.}
}
\value{
a \eqn{2 \times 1} column vector representing the initial coin state.
}
\description{
This returns the chosen initial coin state for the walk.
The supported choices are \code{"right"}, \code{"left"}, and
\code{"symmetric"}.
}
\examples{
init_coin_state("right")
init_coin_state("symmetric")

}
