% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initialize_state}
\alias{initialize_state}
\title{Initialize the full quantum walk state.}
\usage{
initialize_state(T, init_coin)
}
\arguments{
\item{T}{represents the number of walk steps.}

\item{init_coin}{represents the initial coin state used to start the walk.}
}
\value{
a \eqn{2 \times (2T+1)} matrix representing the initial walk state.
}
\description{
This creates the state matrix for the walk, placing the walker
at the origin with the chosen initial coin state.
}
\examples{
initialize_state(2, "right")

}
