% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{kraus_depolarizing}
\alias{kraus_depolarizing}
\title{Construct the Kraus operators for the depolarizing channel.}
\usage{
kraus_depolarizing(p)
}
\arguments{
\item{p}{represents the decoherence strength.}
}
\value{
a list of Kraus operators for the depolarizing channel.
}
\description{
This returns the Kraus operator family for the depolarizing
channel acting on the coin space. The operators are
\deqn{
K_0 = \sqrt{1-p}\,I_c,\qquad
K_1 = \sqrt{p/3}\,X,\qquad
K_2 = \sqrt{p/3}\,Y,\qquad
K_3 = \sqrt{p/3}\,Z.
}
}
\examples{
kraus_depolarizing(0.1)

}
