% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mean_pos}
\alias{mean_pos}
\title{Compute the mean position.}
\usage{
mean_pos(pos, probs)
}
\arguments{
\item{pos}{represents the position grid.}

\item{probs}{represents the probability distribution on the position grid.}
}
\value{
a numeric value representing the mean position.
}
\description{
This computes the mean of a probability distribution on
the position grid. The mean is
\deqn{
\mu_t = \sum_x x\,P_t(x).
}
}
\examples{
mean_pos(c(-1, 0, 1), c(0.25, 0.5, 0.25))

}
