% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{normalize_state}
\alias{normalize_state}
\title{Normalize a walk state.}
\usage{
normalize_state(state, tol = 1e-12)
}
\arguments{
\item{state}{represents the state matrix of the walk.}

\item{tol}{represents the tolerance used to detect a near-zero norm.}
}
\value{
a normalized state matrix.
}
\description{
This rescales the walk state so that its squared norm is 1.
}
\examples{
state <- initialize_state(1, "right")
normalize_state(state)

}
