% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pauli_X}
\alias{pauli_X}
\title{Construct the Pauli X matrix.}
\usage{
pauli_X()
}
\value{
a \eqn{2 \times 2} matrix representing the Pauli X operator.
}
\description{
This returns the Pauli X operator acting on the
two-dimensional coin space.
}
\examples{
pauli_X()

}
