% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pauli_Y}
\alias{pauli_Y}
\title{Construct the Pauli Y matrix.}
\usage{
pauli_Y()
}
\value{
a complex \eqn{2 \times 2} matrix representing the Pauli Y operator.
}
\description{
This returns the Pauli Y operator acting on the
two-dimensional coin space.
}
\examples{
pauli_Y()

}
