% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pauli_Z}
\alias{pauli_Z}
\title{Construct the Pauli Z matrix.}
\usage{
pauli_Z()
}
\value{
a \eqn{2 \times 2} matrix representing the Pauli Z operator.
}
\description{
This returns the Pauli Z operator acting on the
two-dimensional coin space.
}
\examples{
pauli_Z()

}
