% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm_p_value_mc}
\alias{perm_p_value_mc}
\title{Compute Monte Carlo permutation p-value}
\usage{
perm_p_value_mc(observed, perm_stats, alternative = "two.sided")
}
\arguments{
\item{observed}{Observed test statistic.}

\item{perm_stats}{Vector of permuted statistics.}

\item{alternative}{"two.sided", "greater", or "less".}
}
\value{
Numeric p-value.
}
\description{
Computes the Monte Carlo permutation p-value using the standard
(+1)/(B+1) adjustment.
}
\examples{
perm_p_value_mc(2, rnorm(1000), alternative = "greater")

}
