% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm_stat_ks}
\alias{perm_stat_ks}
\title{Compute Kolmogorov–Smirnov (KS) Test statistic}
\usage{
perm_stat_ks(outcome, group)
}
\arguments{
\item{outcome}{A numeric outcome vector.}

\item{group}{A group vector with two levels.}
}
\value{
Numeric KS statistic (maximum absolute difference between ECDFs).
}
\description{
Computes the maximum difference between the Empirical Cumulative Distribution
Functions (ECDFs) of two groups.
}
\examples{
outcome <- c(1, 2, 3, 10, 11, 12)
group <- rep(c("More developed", "Less developed"), each = 3)
perm_stat_mean_diff(outcome, group)

}
