% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perm_dist.R
\name{plot_perm_dist}
\alias{plot_perm_dist}
\title{Plot the Permutation Distribution with Observed Statistic}
\usage{
plot_perm_dist(perm_result, bins = 40)
}
\arguments{
\item{perm_result}{A list containing:
\describe{
\item{perm_stats}{Numeric vector of permutation statistic. }
\item{observed}{Observed test statistic. }
}}

\item{bin}{An Integer specifying the number of histogram bins. Defaults to 40.}
}
\value{
Invisibly returns a list with:
\describe{
\item{hist}{Histogram object returned by function hist().}
\item{obs}{Observed test statistic. }
}
}
\description{
Draws a histogram of the permutation test statistic and shows the observed
test statistic on the histogram as a vertical line.
}
\examples{
perm_res <- list(perm_stats = abs(rnorm(1000)), observed = 2.5)
plot_perm_dist(perm_res)

}
