% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pos_dist}
\alias{pos_dist}
\title{Compute the position distribution of a walk state.}
\usage{
pos_dist(state)
}
\arguments{
\item{state}{represents the state matrix of the walk.}
}
\value{
a numeric vector of position probabilities.
}
\description{
This computes the position probabilities by summing the right
and left coin probabilities at each position. For one trajectory the position distribution is
\deqn{
P_t^{(j)}(x) = |a_t^{(j)}(x)|^2 + |b_t^{(j)}(x)|^2.
}
}
\examples{
state <- initialize_state(2, "right")
pos_dist(state)

}
