% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pos_grid}
\alias{pos_grid}
\title{Create the position grid for the walk.}
\usage{
pos_grid(T)
}
\arguments{
\item{T}{represents the number of walk steps.}
}
\value{
an integer vector of positions from \code{-T} to \code{T}.
}
\description{
This returns the integer position grid from \code{-T} to
\code{T}, which is used for the quantum and classical walks.
}
\examples{
pos_grid(2)

}
