% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{proj_left}
\alias{proj_left}
\title{Construct the projector onto the left coin state.}
\usage{
proj_left()
}
\value{
a \eqn{2 \times 2} projector matrix.
}
\description{
This returns the rank one projector
\eqn{|\mathrm{left}\rangle\langle\mathrm{left}|}.
}
\examples{
proj_left()

}
