% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{proj_right}
\alias{proj_right}
\title{Construct the projector onto the right coin state.}
\usage{
proj_right()
}
\value{
a \eqn{2 \times 2} projector matrix.
}
\description{
This returns the rank one projector
\eqn{|\mathrm{right}\rangle\langle\mathrm{right}|}.
}
\examples{
proj_right()

}
