% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Reference data on reaction time and aging using the Nintendo Wii Balance Board}
\format{
\subsection{\code{rawdata}}{

A data frame with 354 rows and 11 columns
\describe{
\item{testperson_id}{The test subjects ID}
\item{gender}{Male = 1, Female = 2}
\item{age_years}{The test subjects age in years}
\item{medication_count}{The number of medications the test subject is taking}
\item{smoking_status}{1 = Smoker, 2 = Non Smoker, 3 = Previous Smoker}
\item{activiy_work}{Categorical physical activity level at work (1 = Low, 4 = High)}
\item{activiy_sparetime}{Categorical physical activity level in spare time (1 = Low, 4 = High)}
\item{RHD}{Reaction time for dominant hand (ms)}
\item{RHND}{Reaction time for non dominant hand (ms)}
\item{RFD}{Reaction time for dominant foot (ms)}
\item{RFND}{Reaction time for non dominant foot (ms)}
}
}
}
\source{
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0189598}
}
\usage{
rawdata
}
\description{
This is a subset of the data collected for reaction time and aging using the Nintendo Wii Balance Board.
It contains information about the patients age, activity levels, smoking status, and their reaction times
for their dominant and non dominant hands and feet.
}
\keyword{datasets}
