% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_kraus_index}
\alias{sample_kraus_index}
\title{Sample a Kraus index.}
\usage{
sample_kraus_index(probs)
}
\arguments{
\item{probs}{a numeric probability vector.}
}
\value{
an integer giving the sampled Kraus index.
}
\description{
This samples one Kraus operator index according to the
supplied probability vector.
}
\examples{
set.seed(1)
sample_kraus_index(c(0.2, 0.8))

}
