% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_noiseless_qrw}
\alias{sim_noiseless_qrw}
\title{Simulate a noiseless Hadamard quantum random walk.}
\usage{
sim_noiseless_qrw(T, init_coin)
}
\arguments{
\item{T}{represents the number of walk steps.}

\item{init_coin}{represents the initial coin state. The allowed choices are
\code{"right"}, \code{"left"}, and \code{"symmetric"}.}
}
\value{
a list containing the number of steps, the time indexed position
distributions, the mean positions, and the variances.
}
\description{
This simulates the deterministic Hadamard quantum random walk
on a line. The evolution is generated by repeated application of
\deqn{
U = S(H \otimes I_p).
}
}
\examples{
sim_noiseless_qrw(T = 3, init_coin = "symmetric")

}
