% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_srw}
\alias{sim_srw}
\title{Simulate a symmetric classical random walk.}
\usage{
sim_srw(T)
}
\arguments{
\item{T}{represents the number of walk steps.}
}
\value{
a list containing the number of steps, the time indexed position
distributions, the mean positions, and the variances.
}
\description{
This computes the exact position distributions, mean positions,
and variances for a symmetric classical random walk on a line from time
\code{0} to time \code{T}. For the symmetric random walk,
\eqn{\mathbb{E}[X_t] = 0} and \eqn{\mathrm{Var}(X_t) = t}.
}
\examples{
sim_srw(3)

}
