% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{state_norm_sq}
\alias{state_norm_sq}
\title{Compute the squared norm of a walk state.}
\usage{
state_norm_sq(state)
}
\arguments{
\item{state}{represents the state matrix of the walk.}
}
\value{
a numeric value equal to the squared norm of the state.
}
\description{
This computes the total probability mass of the walk state by
summing the squared moduli of all amplitudes.
}
\examples{
state <- initialize_state(1, "right")
state_norm_sq(state)

}
