% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{var_pos}
\alias{var_pos}
\title{Compute the position variance.}
\usage{
var_pos(pos, probs)
}
\arguments{
\item{pos}{represents the position grid.}

\item{probs}{represents the probability distribution on the position grid.}
}
\value{
a numeric value representing the position variance.
}
\description{
This computes the variance of a probability distribution on
the position grid. The variance is
\deqn{
\mathrm{Var}(X_t) = \sum_x (x - \mu_t)^2 P_t(x).
}
}
\examples{
var_pos(c(-1, 0, 1), c(0.25, 0.5, 0.25))

}
