% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{get_clean_numeric}
\alias{get_clean_numeric}
\title{Removes decimal points, strings and NA values from a vector.}
\usage{
get_clean_numeric(data)
}
\arguments{
\item{data}{the vector to clean.}
}
\value{
A vector containing the cleaned data.
}
\description{
Removes decimal points, characters, strings, and missing values from an inputted vector.
}
\examples{
x <- c("1,000", "2,500", "", NA, 1)
get_clean_numeric(x)
}
